refactor(seperate service for youtube db): seperate service for youtube db
seperate service for youtube db
